+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+#### 1.0.0+1 - 2017-09-07
+  * Strong-mode fix in InfiniteIterable.fold(), from @giorgiofran.
+
 #### 1.0.0-dev.1
-   * Split from `quiver-dart`
+  * Split from `quiver-dart`